# assignment-1-416C616E/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is my(Alan Neads) OpenAPI documentation for the yelp-like review site we were tasked with making for the first assignment for CS 493, Cloud Application Development.",
    "version": "1.0.0",
    "title": "Business Review Site",
    "termsOfService": "",
    "contact": {
      "email": "neadsa@oregonstate.edu"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/user/{userID}/businesses": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get a list of the users' businesses",
        "description": "Get a list of the users' businesses",
        "operationId": "getUserBusinesses",
        "parameters": [
          {
            "name": "userID",
            "in": "path",
            "description": "ID of user",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Businesses returned as response",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Business"
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/user/{userID}/reviews": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get a list of the users' reviews",
        "description": "Get a list of the users' reviews",
        "operationId": "getUserReviews",
        "parameters": [
          {
            "name": "userID",
            "in": "path",
            "description": "ID of user",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Reviews returned as response",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Review"
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Category": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Category"
      }
    },
    "Subcategory": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "categoryID": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Subcategory"
      }
    },
    "Business": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "ownerUserID": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "streetAddress": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "state": {
          "type": "string"
        },
        "zipCode": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "categories": {
          "type": "array",
          "description": "List referencing categories",
          "items": {
            "$ref": "#/definitions/Category"
          }
        },
        "subcategories": {
          "type": "array",
          "description": "List referencing subcategories",
          "items": {
            "$ref": "#/definitions/Subcategory"
          }
        },
        "website": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Business"
      }
    },
...
    "Review": {
      "type": "object",
      "properties": {
        "businessID": {
          "type": "integer",
          "format": "int64"
        },
        "userID": {
          "type": "integer",
          "format": "int64"
        },
        "starRating": {
          "type": "integer",
          "format": "int64"
        },
        "dollarSignRating": {
          "type": "integer",
          "format": "int64"
        },
        "writtenReview": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Image"
      }
    }
  }
}